+++ src/hayate/middleware/__init__.py
@@ -9,11 +9,14 @@ from .compress import compress
 from .cors import cors
 from .etag import etag
 from .logger import logger
+from .rate_limit import MemoryRateLimitStore, RateLimitStore, rate_limit
 from .secure_headers import secure_headers
 from .static_files import static_files
 from .timeout import timeout
 
 __all__ = [
+    "MemoryRateLimitStore",
+    "RateLimitStore",
     "basic_auth",
     "body_limit",
     "cache",
@@ -21,6 +24,7 @@ __all__ = [
     "cors",
     "etag",
     "logger",
+    "rate_limit",
     "secure_headers",
     "static_files",
     "timeout",